Update DictionaryList.cpp
Changed the find() function to find_first() since it looks for the first occurrence of a value. 
Added a regular and constant find() function to search inside a list by memory address rather than by value. 
The print() function has been changed to work correctly with a list of <int*> pointers.